Work around GitHub bug with URLs containing #
Apparently the GitHub API doesn't properly encode the URLs it returns
via JSON: the first # character in each URL always remains unencoded.
In the rare (yet encountered) cases of file containing # characters,
this leads to entirely avoidable 404s.

(I think this is related to fragment identifiers, and it's clearly a
bug in whatever URL handling or JSON library GitHub uses.)