Prevent "over-sampling" on the second query
If the population of a stratum is between 1000 and 2000, we were
re-sampling files during the second ("reverse") query that we'd already
seen during the first query. This isn't really a problem, as each file
is only stored once in the database, but it screws up the sampling
statistics, and it's also an unnecessary waste of time. So from now on
we just stop sampling once we've reached the population count.

There is a risk now of missing some files that we would've otherwise
gotten (due to the population count underestimating the number of files
actually returned), but I think it's pretty small, especially since the
population count is the conservative maximum of all the population
counts the API has given us for the stratum, which is usually
relatively stable for the amount of files we are talking about here.